moved get_model,  set_seed. Depricated etnn/utils
