+++ main_qm9.py
@@ -1,7 +1,4 @@
-# import argparse
 import copy
-
-# import hashlib
 import json
 import logging
 import os
@@ -10,19 +7,13 @@ import time
 
 import hydra
 import torch
-
-# import parser_utils
 import wandb
 from omegaconf import DictConfig, OmegaConf
 from torch_geometric.loader import DataLoader
 from tqdm import tqdm
 
 from etnn.qm9.qm9cc import QM9CC
-
-from utils import args_to_hash, calc_mean_mad
-
-# from src.utils import get_loaders, get_model, set_seed
-from etnn.utils import get_model, set_seed
+from utils import args_to_hash, calc_mean_mad, get_model, set_seed
 
 torch.set_float32_matmul_precision("high")
 os.environ["WANDB__SERVICE_WAIT"] = "600"
